Update demo links and repository clone instructions
Replaced outdated links with new ones for the live demo and image thumbnail image. This ensures users access the most recent resources and improves clarity on how to clone the repository correctly. These updates help maintain accurate documentation and create a smoother onboarding experience for new users.

Fix image link format in README

Updated the image link in the README to use the correct format, ensuring direct access to the image file. This improves the clarity and functionality of the documentation by making the image readily available for viewers, enhancing the overall user experience.

Update image thumbnail in README and add new photo

Replaced the existing thumbnail with a more relevant image to enhance the visual representation of the project. This update helps to better align the README with the project's purpose and improves overall presentation.

Improve README layout and clarify demo access

Updated the README to enhance clarity and organization by repositioning the image thumbnail and refining the introductory message. This improves the visual appeal and better guides users to the Live Demo, ensuring a smoother onboarding experience for new users engaging with the image roasting feature.